http: move utcDate to internal/http.js
