+++ lib/_http_outgoing.js
@@ -23,9 +23,9 @@
 
 const assert = require('assert').ok;
 const Stream = require('stream');
-const timers = require('timers');
 const util = require('util');
 const internalUtil = require('internal/util');
+const internalHttp = require('internal/http');
 const Buffer = require('buffer').Buffer;
 const common = require('_http_common');
 const checkIsHttpToken = common._checkIsHttpToken;
@@ -37,6 +37,7 @@ const errors = require('internal/errors');
 
 const CRLF = common.CRLF;
 const debug = common.debug;
+const utcDate = internalHttp.utcDate;
 
 var RE_FIELDS =
   /^(?:Connection|Transfer-Encoding|Content-Length|Date|Expect|Trailer|Upgrade)$/i;
@@ -64,23 +65,7 @@ function isCookieField(s) {
   return true;
 }
 
-var dateCache;
-function utcDate() {
-  if (!dateCache) {
-    var d = new Date();
-    dateCache = d.toUTCString();
-    timers.enroll(utcDate, 1000 - d.getMilliseconds());
-    timers._unrefActive(utcDate);
-  }
-  return dateCache;
-}
-utcDate._onTimeout = function _onTimeout() {
-  dateCache = undefined;
-};
-
-
 function noopPendingOutput(amount) {}
-
 
 function OutgoingMessage() {
   Stream.call(this);
